--- trakker/src/main/java/com/example/trakker/service/heart/HeartService.java
@@ -1,7 +1,0 @@
-package com.example.trakker.service.heart;
-
-import com.example.trakker.item.HeartDTO;
-
-public interface HeartService {
-    void heartCheck(HeartDTO dto);
-}
