+++ trakker/src/main/java/com/example/trakker/service/item/HeartService.java
@@ -1,0 +1,7 @@
+package com.example.trakker.service.item;
+
+import com.example.trakker.item.HeartDTO;
+
+public interface HeartService {
+    void heartCheck(HeartDTO dto);
+}
