+++ trakker/src/main/java/com/example/trakker/service/item/LocalService.java
@@ -1,0 +1,13 @@
+package com.example.trakker.service.item;
+
+import com.example.trakker.item.LocalDTO;
+
+import java.util.List;
+
+public interface LocalService {
+    int localCount();
+
+    String selectLocal(Integer lNum);
+
+    List<LocalDTO> localList();
+}
